requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-google-genai
langchain-community
langchain-core
langchain-text-splitters
langgraph
beautifulsoup4